Fix gallery blowing up when importing gallery part
On the `GalleryPart` the `MediaItems` are currently being stored as a
JSON string. The issue this caused was when importing it wouldn't
import as a list of items, which caused a syntax error when trying to
parse the JSON.

To fix this, and improve the storage in general, the `MediaItems`
property has been changed to an typed array containing concrete objects
that represent the gallery items.