Split image types in to external/internal
Currently there is no way to differentiate between images that are
via the media library or via a direct URL. The issue with this is we may
want to render the output of images via the media library differently
in order to make use of the image resizer.

This update also fixes #11 as instead of storing the full URL to media
added via the media library, only the path to the image is stored. This
will prevent images becoming broken when using gallery on tenant with
a URL prefix, as currently the URL prefix is stored within the data.