Fix issue with images not showing when editing
When adding a new image to the gallery, it wouldn't display until the
gallery was published. This was due to a casing issue with the JSON
properties used for the data. To solve this a view model class has
been added which uses `JsonProperty` to format the data.

This actually highlighted another issue with the thumbnail being stored
within the data. This property has been removed and is generated within
the display driver.